use server safe popover for image alt text in blog
